Create separate packages for activator test bundles.
- When running tests in IntelliJ, classes from the two bundles
  are confused.
