+++ jdisc_core_test/integration_test/src/test/java/com/yahoo/jdisc/application/BundleActivatorIntegrationTest.java
@@ -22,7 +22,7 @@ public class BundleActivatorIntegrationTest {
         OsgiFramework osgi = driver.osgiFramework();
         BundleInstaller installer = new BundleInstaller(driver.osgiFramework());
         Bundle bundle = installer.installAndStart("my-bundle-activator.jar").get(0);
-        Class<?> serviceClass = bundle.loadClass("com.yahoo.jdisc.bundle.MyService");
+        Class<?> serviceClass = bundle.loadClass("com.yahoo.jdisc.bundle.my_act.MyService");
         assertNotNull(serviceClass);
         BundleContext ctx = osgi.bundleContext();
         ServiceReference<?> serviceRef = ctx.getServiceReference(serviceClass.getName());
@@ -37,7 +37,7 @@ public class BundleActivatorIntegrationTest {
     public void requireThatApplicationBundleActivatorHasAccessToCurrentContainer() throws Exception {
         TestDriver driver = TestDriver.newApplicationBundleInstance("app-g-act.jar", false);
         OsgiFramework osgi = driver.osgiFramework();
-        Class<?> serviceClass = osgi.bundles().get(1).loadClass("com.yahoo.jdisc.bundle.MyService");
+        Class<?> serviceClass = osgi.bundles().get(1).loadClass("com.yahoo.jdisc.bundle.g_act.MyService");
         assertNotNull(serviceClass);
         BundleContext ctx = osgi.bundleContext();
         ServiceReference<?> serviceRef = ctx.getServiceReference(serviceClass.getName());
