+++ jdisc_core_test/test_bundles/my-bundle-activator/src/main/java/com/yahoo/jdisc/bundle/my_act/MyBundleActivator.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.jdisc.bundle;
+package com.yahoo.jdisc.bundle.my_act;
 
 import com.yahoo.jdisc.service.CurrentContainer;
 import org.osgi.framework.BundleActivator;
